backend/requirements.txt
amqp==5.1.1
asgiref==3.7.2
billiard==4.1.0
braintree==4.20.0
Brotli==1.0.9
celery==5.3.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
cssselect2==0.7.0
Django==4.2.2
django-localflavor==4.0
django-rosetta==0.9.9
djangorestframework==3.14.0
flower==1.2.0
fonttools==4.40.0
gobject==0.1.0
html5lib==1.1
humanize==4.6.0
idna==3.4
kombu==5.3.0
Pillow==9.5.0
polib==1.2.0
prometheus-client==0.17.0
prompt-toolkit==3.0.38
pycparser==2.21
pydyf==0.6.0
Pygments==2.15.1
pyphen==0.14.0
python-dateutil==2.8.2
python-decouple==3.8
python-gettext==5.0
python-stdnum==1.18
pytz==2023.3
redis==4.5.5
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tinycss2==1.2.1
tornado==6.3.2
tzdata==2023.3
urllib3==2.0.3
vine==5.0.0
wcwidth==0.2.6
weasyprint==59.0
webencodings==0.5.1
zopfli==0.2.2
